refactor: command hooks as interceptors
