+++ cmd/aspect/root/BUILD.bazel
@@ -18,6 +18,7 @@ go_library(
         "//cmd/aspect/test",
         "//cmd/aspect/version",
         "//docs/help/topics",
+        "//pkg/aspect/root/flags",
         "//pkg/ioutils",
         "//pkg/plugin/system",
         "@com_github_fatih_color//:color",
