+++ cmd/aspect/test/BUILD.bazel
@@ -6,9 +6,7 @@ go_library(
     importpath = "aspect.build/cli/cmd/aspect/test",
     visibility = ["//visibility:public"],
     deps = [
-        "//cmd/aspect/root/flags",
         "//pkg/aspect/test",
-        "//pkg/aspecterrors",
         "//pkg/bazel",
         "//pkg/interceptors",
         "//pkg/ioutils",
