+++ pkg/aspect/root/flags/BUILD.bazel
@@ -3,6 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "flags",
     srcs = ["config.go"],
-    importpath = "aspect.build/cli/cmd/aspect/root/flags",
-    visibility = ["//visibility:public"],
+    importpath = "aspect.build/cli/pkg/aspect/root/flags",
+    visibility = ["//:__subpackages__"],
 )
